feat(viewer): create textured IFC objects from visual scan regions (#4399)
* docs(appearance): define captured mesh integration and fidelity gates (#4380)

* refactor(viewer): share textured product authoring transaction

* feat(viewer): connect captured meshes to shared IFC authoring transaction

* feat(viewer): resolve retained capture images and share authoring destination defaults

* test(viewer): exercise captured creation transaction and own source snapshot (#4380)

* test(viewer): skip native capture contract when WASM runtime is absent

* feat(viewer): extract registered captured triangle regions without losing UV seams

* fix(authoring): preserve captured image sampler flags (#4380)

* fix(viewer): preserve captured image repeat samplers for native authoring

* feat(cache): preserve captured GLB albedo and source images (#4380)

* fix(viewer): preserve captured image repeat samplers for native authoring

* fix(authoring): preserve captured image sampler flags (#4380)

* feat(viewer): bind captured regions to registered source images and samplers

* fix(viewer): bound captured image validation and prove load ownership

* test(viewer): retain IFC schema literal in capture fixture

* fix(viewer): bound capture rows and offer only editable IFC destinations (#4380)

* feat(viewer): select textured scan regions and create IFC objects in Appearance

* fix(viewer): guard capture preview lifetime and refresh registered selections

* docs(viewer): record complete and regional captured IFC acceptance

* docs: verify captured IFC appearance through fresh room sharing (#4380)

* docs: verify selected captured IFCZIP retains texture and containment (#4380)

* fix(viewer): retain explicit capture source choice after removal (#4380)

+++ apps/viewer/src/lib/appearance/capture/region.ts
@@ -60,7 +60,7 @@ export function captureRegion(mesh: MeshData, triangleIds: readonly number[], re
   registration.validate();
   if (registration.source !== mesh) throw new Error('The registration belongs to another captured surface.');
   if (!mesh.textureRef || mesh.texture) throw new Error('The captured surface needs one external image texture.');
-  if (mesh.color.some(value => value !== 1)) throw new Error('Bake the material tint and opacity into the image before IFC capture.');
+  if (mesh.color.some(value => Math.fround(value) !== 1)) throw new Error('Bake the material tint and opacity into the image before IFC capture.');
   if ((mesh.geometryClass ?? 0) === 2 || mesh.entityIds) throw new Error('Choose one concrete surface occurrence.');
   if (triangleIds.length === 0 || triangleIds.length > MAX_ROWS) throw new Error('Choose 1..200000 captured triangles.');
   const vertexCount = mesh.positions.length / 3;
